+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- `Graph` - Added context menus for nodes, edges and the graph background with relevant actions.
 - `ElementsInfoPanel` - Added a new panel to display some information about the currently selected nodes and edges.
 - `Algorithms` - Added traversal algorithms tab in the Algorithms modal.
 - `Project` - Adopted SonarQube for code quality analysis.